+++ src/main/java/com/sprint/mission/discodeit/exception/advice/AuthExceptionHandler.java
@@ -1,7 +1,9 @@
-package com.sprint.mission.discodeit.exception.channel;
+package com.sprint.mission.discodeit.exception.advice;
 
-import com.sprint.mission.discodeit.controller.ChannelController;
+import com.sprint.mission.discodeit.controller.AuthController;
 import com.sprint.mission.discodeit.dto.response.ErrorResponse;
+import com.sprint.mission.discodeit.exception.auth.LoginFailedException;
+import com.sprint.mission.discodeit.exception.auth.UsernameNotFoundException;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
@@ -9,27 +11,11 @@ import org.springframework.web.bind.annotation.ExceptionHandler;
 import org.springframework.web.bind.annotation.RestControllerAdvice;
 
 @Slf4j
-@RestControllerAdvice(basePackageClasses = ChannelController.class)
-public class ChannelExceptionHandler {
-    @ExceptionHandler(ChannelNotFoundException.class)
-    public ResponseEntity<ErrorResponse> handleException(ChannelNotFoundException e) {
-        log.error("존재하지 않는 채널 - 에러 메시지: {}", e.getMessage());
-        ErrorResponse result = new ErrorResponse(
-                e.getTimestamp(),
-                e.getErrorCode().toString(),
-                e.getMessage(),
-                e.getDetails(),
-                e.getClass().getName(),
-                404
-        );
-        return ResponseEntity
-                .status(HttpStatus.NOT_FOUND)
-                .body(result);
-    }
-
-    @ExceptionHandler(PrivateChannelUpdateException.class)
-    public ResponseEntity<ErrorResponse> handleException(PrivateChannelUpdateException e) {
-        log.error("비공개 채널 수정 불가 - 에러 메시지: {}", e.getMessage());
+@RestControllerAdvice(basePackageClasses = AuthController.class)
+public class AuthExceptionHandler {
+    @ExceptionHandler(LoginFailedException.class)
+    public ResponseEntity<ErrorResponse> handleException(LoginFailedException e) {
+        log.error("로그인 실패 - 에러 메시지: {}", e.getMessage());
         ErrorResponse result = new ErrorResponse(
                 e.getTimestamp(),
                 e.getErrorCode().toString(),
@@ -40,6 +26,22 @@ public class ChannelExceptionHandler {
         );
         return ResponseEntity
                 .status(HttpStatus.BAD_REQUEST)
+                .body(result);
+    }
+
+    @ExceptionHandler(UsernameNotFoundException.class)
+    public ResponseEntity<ErrorResponse> handleException(UsernameNotFoundException e) {
+        log.error("존재하지 않는 사용자명 - 에러 메시지: {}", e.getMessage());
+        ErrorResponse result = new ErrorResponse(
+                e.getTimestamp(),
+                e.getErrorCode().toString(),
+                e.getMessage(),
+                e.getDetails(),
+                e.getClass().getName(),
+                404
+        );
+        return ResponseEntity
+                .status(HttpStatus.NOT_FOUND)
                 .body(result);
     }
 }
