+++ src/main/java/com/sprint/mission/discodeit/exception/advice/ChannelExceptionHandler.java
@@ -1,7 +1,9 @@
-package com.sprint.mission.discodeit.exception.userstatus;
+package com.sprint.mission.discodeit.exception.advice;
 
-import com.sprint.mission.discodeit.controller.UserController;
+import com.sprint.mission.discodeit.controller.ChannelController;
 import com.sprint.mission.discodeit.dto.response.ErrorResponse;
+import com.sprint.mission.discodeit.exception.channel.ChannelNotFoundException;
+import com.sprint.mission.discodeit.exception.channel.PrivateChannelUpdateException;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
@@ -9,11 +11,11 @@ import org.springframework.web.bind.annotation.ExceptionHandler;
 import org.springframework.web.bind.annotation.RestControllerAdvice;
 
 @Slf4j
-@RestControllerAdvice(basePackageClasses = UserController.class)
-public class UserStatusExceptionHandler {
-    @ExceptionHandler(UserStatusNotFoundException.class)
-    public ResponseEntity<ErrorResponse> handleException(UserStatusNotFoundException e) {
-        log.error("존재하지 않는 사용자 상태 - 에러 메시지: {}", e.getMessage());
+@RestControllerAdvice(basePackageClasses = ChannelController.class)
+public class ChannelExceptionHandler {
+    @ExceptionHandler(ChannelNotFoundException.class)
+    public ResponseEntity<ErrorResponse> handleException(ChannelNotFoundException e) {
+        log.error("존재하지 않는 채널 - 에러 메시지: {}", e.getMessage());
         ErrorResponse result = new ErrorResponse(
                 e.getTimestamp(),
                 e.getErrorCode().toString(),
@@ -27,9 +29,9 @@ public class UserStatusExceptionHandler {
                 .body(result);
     }
 
-    @ExceptionHandler(UserStatusDuplicateException.class)
-    public ResponseEntity<ErrorResponse> handleException(UserStatusDuplicateException e) {
-        log.error("사용자 상태 중복 생성 불가 - 에러 메시지: {}", e.getMessage());
+    @ExceptionHandler(PrivateChannelUpdateException.class)
+    public ResponseEntity<ErrorResponse> handleException(PrivateChannelUpdateException e) {
+        log.error("비공개 채널 수정 불가 - 에러 메시지: {}", e.getMessage());
         ErrorResponse result = new ErrorResponse(
                 e.getTimestamp(),
                 e.getErrorCode().toString(),
