+++ src/main/java/com/sprint/mission/discodeit/exception/advice/MessageExceptionHandler.java
@@ -1,7 +1,8 @@
-package com.sprint.mission.discodeit.exception.binarycontent;
+package com.sprint.mission.discodeit.exception.advice;
 
-import com.sprint.mission.discodeit.controller.BinaryContentController;
+import com.sprint.mission.discodeit.controller.MessageController;
 import com.sprint.mission.discodeit.dto.response.ErrorResponse;
+import com.sprint.mission.discodeit.exception.message.MessageNotFoundException;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
@@ -9,11 +10,11 @@ import org.springframework.web.bind.annotation.ExceptionHandler;
 import org.springframework.web.bind.annotation.RestControllerAdvice;
 
 @Slf4j
-@RestControllerAdvice(basePackageClasses = BinaryContentController.class)
-public class BinaryContentExceptionHandler {
-    @ExceptionHandler(BinaryContentNotFoundException.class)
-    public ResponseEntity<ErrorResponse> handleException(BinaryContentNotFoundException e) {
-        log.error("존재하지 않는 파일 - 에러 메시지: {}", e.getMessage());
+@RestControllerAdvice(basePackageClasses = MessageController.class)
+public class MessageExceptionHandler {
+    @ExceptionHandler(MessageNotFoundException.class)
+    public ResponseEntity<ErrorResponse> handleException(MessageNotFoundException e) {
+        log.error("존재하지 않는 메시지 - 에러 메시지: {}", e.getMessage());
         ErrorResponse result = new ErrorResponse(
                 e.getTimestamp(),
                 e.getErrorCode().toString(),
