--- src/main/java/com/sprint/mission/discodeit/exception/auth/AuthExceptionHandler.java
@@ -1,45 +1,0 @@
-package com.sprint.mission.discodeit.exception.auth;
-
-import com.sprint.mission.discodeit.controller.AuthController;
-import com.sprint.mission.discodeit.dto.response.ErrorResponse;
-import lombok.extern.slf4j.Slf4j;
-import org.springframework.http.HttpStatus;
-import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.ExceptionHandler;
-import org.springframework.web.bind.annotation.RestControllerAdvice;
-
-@Slf4j
-@RestControllerAdvice(basePackageClasses = AuthController.class)
-public class AuthExceptionHandler {
-    @ExceptionHandler(LoginFailedException.class)
-    public ResponseEntity<ErrorResponse> handleException(LoginFailedException e) {
-        log.error("로그인 실패 - 에러 메시지: {}", e.getMessage());
-        ErrorResponse result = new ErrorResponse(
-                e.getTimestamp(),
-                e.getErrorCode().toString(),
-                e.getMessage(),
-                e.getDetails(),
-                e.getClass().getName(),
-                400
-        );
-        return ResponseEntity
-                .status(HttpStatus.BAD_REQUEST)
-                .body(result);
-    }
-
-    @ExceptionHandler(UsernameNotFoundException.class)
-    public ResponseEntity<ErrorResponse> handleException(UsernameNotFoundException e) {
-        log.error("존재하지 않는 사용자명 - 에러 메시지: {}", e.getMessage());
-        ErrorResponse result = new ErrorResponse(
-                e.getTimestamp(),
-                e.getErrorCode().toString(),
-                e.getMessage(),
-                e.getDetails(),
-                e.getClass().getName(),
-                404
-        );
-        return ResponseEntity
-                .status(HttpStatus.NOT_FOUND)
-                .body(result);
-    }
-}
